Refactor SpringConfiguration to extend from new BaseConfiguration; extend BaseAppConfiguration from BaseConfiguration; configure DataSources upon creation; refactor SpringConfiguration to use common BaseConfiguration functions
