+++ apps/bfd-shared-utils/src/main/java/gov/cms/bfd/sharedutils/database/DataSourceFactory.java
@@ -1,5 +1,6 @@
 package gov.cms.bfd.sharedutils.database;
 
+import com.codahale.metrics.MetricRegistry;
 import javax.sql.DataSource;
 
 /** Implementations of this interface create {@link DataSource} instances on demand. */
@@ -11,4 +12,14 @@ public interface DataSourceFactory {
    * @return the instance
    */
   DataSource createDataSource();
+
+  /**
+   * Create a properly configured {@link DataSource} instance that is configured with the provided
+   * {@link MetricRegistry}.
+   *
+   * @param metricRegistry the {@link MetricRegistry} that will be used to generate metrics on the
+   *     {@link DataSource}
+   * @return the instance
+   */
+  DataSource createDataSource(MetricRegistry metricRegistry);
 }
