+++ apps/bfd-shared-utils/src/main/java/gov/cms/bfd/sharedutils/database/DefaultHikariDataSourceFactory.java
@@ -1,6 +1,8 @@
 package gov.cms.bfd.sharedutils.database;
 
+import com.codahale.metrics.MetricRegistry;
 import com.zaxxer.hikari.HikariDataSource;
+import javax.annotation.Nullable;
 import lombok.AllArgsConstructor;
 
 /**
@@ -14,8 +16,13 @@ public class DefaultHikariDataSourceFactory implements HikariDataSourceFactory {
 
   @Override
   public HikariDataSource createDataSource() {
+    return createDataSource(null);
+  }
+
+  @Override
+  public HikariDataSource createDataSource(MetricRegistry metricRegistry) {
     HikariDataSource pooledDataSource = new HikariDataSource();
-    configureDataSource(pooledDataSource);
+    configureDataSource(pooledDataSource, metricRegistry);
     return pooledDataSource;
   }
 
@@ -24,12 +31,34 @@ public class DefaultHikariDataSourceFactory implements HikariDataSourceFactory {
    * source.
    *
    * @param dataSource data source to configure
+   * @param metricRegistry the {@link MetricRegistry} that will be used to generate metrics on the
+   *     {@link HikariDataSource}
    */
-  protected void configureDataSource(HikariDataSource dataSource) {
+  protected void configureDataSource(
+      HikariDataSource dataSource, @Nullable MetricRegistry metricRegistry) {
     dataSource.setJdbcUrl(dbOptions.getDatabaseUrl());
     dataSource.setUsername(dbOptions.getDatabaseUsername());
     dataSource.setPassword(dbOptions.getDatabasePassword());
     dataSource.setMaximumPoolSize(Math.max(2, dbOptions.getMaxPoolSize()));
     dataSource.setRegisterMbeans(true);
+
+    /*
+     * FIXME Temporary workaround for CBBI-357: send Postgres' query planner a
+     * strongly worded letter instructing it to avoid sequential scans whenever
+     * possible.
+     */
+    if (dataSource.getJdbcUrl() != null && dataSource.getJdbcUrl().contains("postgre"))
+      dataSource.setConnectionInitSql(
+          "set application_name = 'bfd-server'; set enable_seqscan = false;");
+
+    if (metricRegistry != null) {
+      dataSource.setMetricRegistry(metricRegistry);
+    }
+
+    /*
+     * FIXME Temporary setting for BB-1233 to find the source of any possible leaks
+     * (see: https://github.com/brettwooldridge/HikariCP/issues/1111)
+     */
+    dataSource.setLeakDetectionThreshold(60 * 1000);
   }
 }
