+++ apps/bfd-shared-utils/src/main/java/gov/cms/bfd/sharedutils/database/HikariDataSourceFactory.java
@@ -1,5 +1,6 @@
 package gov.cms.bfd.sharedutils.database;
 
+import com.codahale.metrics.MetricRegistry;
 import com.zaxxer.hikari.HikariDataSource;
 
 /** Implementations of this interface create {@link HikariDataSource} instances on demand. */
@@ -10,4 +11,14 @@ public interface HikariDataSourceFactory extends DataSourceFactory {
    * @return the instance
    */
   HikariDataSource createDataSource();
+
+  /**
+   * Create a properly configured {@link HikariDataSource} instance that is configured with the
+   * provided {@link MetricRegistry}.
+   *
+   * @param metricRegistry the {@link MetricRegistry} that will be used to generate metrics on the
+   *     {@link HikariDataSource}
+   * @return the instance
+   */
+  HikariDataSource createDataSource(MetricRegistry metricRegistry);
 }
